refactor(vanilla-extract): mv getBoxProps/createBoxSprinkles
