+++ examples/react-basic/package.json
@@ -13,47 +13,30 @@
         "preview": "vite preview",
         "format": "prettier --write \"./{src,tests}/**/*\"",
         "lint:ts": "tsc --noEmit",
-        "lint": "eslint './{src,tests}/**/*.ts' --cache ",
-        "lint:fix": "eslint './{src,tests}/**/*.ts' --cache --fix",
-        "lint:debug": "TIMING=1 eslint './{src,tests}/**/*.ts'"
+        "lint": "../../node_modules/.bin/eslint './src/**/*.ts' --cache --format=pretty",
+        "lint:fix": "../../node_modules/.bin/eslint './src/**/*.ts' --cache --format=pretty --fix"
     },
     "dependencies": {
         "@box-extractor/core": "workspace:*",
+        "@box-extractor/vanilla-extract": "workspace:^0.0.1",
         "@dessert-box/react": "^0.4.0",
-        "@unocss/preset-icons": "^0.46.4",
         "@vanilla-extract/css": "^1.9.2",
-        "@vanilla-extract/recipes": "^0.2.5",
         "@vanilla-extract/sprinkles": "^1.5.0",
-        "eval": "^0.1.8",
-        "javascript-stringify": "^2.1.0",
         "pastable": "^2.0.10",
-        "rainbow-sprinkles": "^0.14.3",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
         "react-error-boundary": "^3.1.4",
-        "ts-morph": "^16.0.0",
-        "ts-pattern": "^4.0.5",
-        "unocss": "^0.46.4"
+        "ts-pattern": "^4.0.5"
     },
     "devDependencies": {
         "@astahmer/eslint-config-react": "^0.1.5",
-        "@iconify/json": "^2.1.135",
         "@types/react": "^18.0.15",
         "@types/react-dom": "^18.0.6",
-        "@typescript-eslint/eslint-plugin": "^5.42.1",
-        "@vanilla-extract/integration": "^6.0.1",
-        "@vanilla-extract/vite-plugin": "link:/Users/astahmer/dev/open-source/vanilla-extract/packages/vite-plugin",
         "@vitejs/plugin-react": "^2.0.0",
         "eslint": "^8.27.0",
-        "eslint-formatter-pretty": "^4.1.0",
-        "fast-glob": "^3.2.12",
         "prettier": "^2.7.1",
         "ts-toolbelt": "^9.6.0",
-        "type-fest": "^3.2.0",
         "typescript": "^4.7.4",
-        "vite": "^3.0.4",
-        "vite-plugin-checker": "^0.4.9",
-        "vite-plugin-compress": "^2.1.1",
-        "vite-plugin-inspect": "^0.7.7"
+        "vite": "^3.0.4"
     }
 }
