+++ packages/adapter-vanilla-extract/package.json
@@ -1,13 +1,13 @@
 {
-    "name": "@box-extractor/core",
+    "name": "@box-extractor/vanilla-extract",
     "version": "0.0.1",
+    "main": "dist/box-extractor-vanilla-extract.cjs.js",
+    "module": "dist/box-extractor-vanilla-extract.esm.js",
     "repository": {
         "type": "git",
         "url": "https://github.com/astahmer/vite-box-extractor.git"
     },
     "author": "Alexandre Stahmer <alexandre.stahmer@gmail.com>",
-    "main": "dist/box-extractor-core.cjs.js",
-    "module": "dist/box-extractor-core.esm.js",
     "scripts": {
         "test": "vitest",
         "lint:ts": "tsc --noEmit",
@@ -16,25 +16,16 @@
     },
     "license": "ISC",
     "dependencies": {
-        "@box-extractor/vanilla-extract": "workspace:*",
-        "@phenomnomnominal/tsquery": "^5.0.0",
+        "@vanilla-extract/css": "^1.9.2",
         "@vanilla-extract/integration": "link:/Users/astahmer/dev/open-source/vanilla-extract/packages/integration",
+        "@vanilla-extract/sprinkles": "link:/Users/astahmer/dev/open-source/vanilla-extract/packages/sprinkles",
         "@vanilla-extract/vite-plugin": "link:/Users/astahmer/dev/open-source/vanilla-extract/packages/vite-plugin",
-        "eval": "^0.1.8",
-        "javascript-stringify": "^2.1.0",
-        "microdiff": "^1.3.1",
-        "pastable": "^2.0.10",
-        "ts-evaluator": "^1.0.7",
-        "ts-morph": "^16.0.0",
-        "ts-pattern": "^4.0.5"
+        "pastable": "^2.0.10"
     },
     "peerDependencies": {
         "vite": "^3.2.3"
     },
     "devDependencies": {
-        "@types/react": "^18.0.15",
-        "@types/react-dom": "^18.0.6",
-        "@vanilla-extract/sprinkles": "link:/Users/astahmer/dev/open-source/vanilla-extract/packages/sprinkles",
         "prettier": "^2.7.1",
         "ts-toolbelt": "^9.6.0",
         "typescript": "^4.7.4",
@@ -43,12 +34,14 @@
     },
     "preconstruct": {
         "entrypoints": [
-            "index.ts"
+            "index.ts",
+            "createRuntimeBoxSprinkles.ts"
         ]
     },
     "files": [
         "src",
         "dist",
+        "createRuntimeBoxSprinkles",
         "../README.md"
     ],
     "keywords": [
