+++ packages/adapter-vanilla-extract/src/index.ts
@@ -1,0 +1,4 @@
+export { createBoxSprinkles } from "./createBoxSprinkles";
+export type { AnySprinklesFn, SprinkleConditionsNames } from "./createBoxSprinklesInternal";
+export { createBoxSprinklesInternal } from "./createBoxSprinklesInternal";
+export { getBoxProps } from "./getBoxProps";
