+++ packages/adapter-vanilla-extract/tsconfig.json
@@ -1,0 +1,12 @@
+{
+    "extends": "../../tsconfig.json",
+    "compilerOptions": {
+        "isolatedModules": true, // https://preconstruct.tools/guides/building-typescript-packages
+        "noEmit": true,
+        "jsx": "react-jsx",
+        "types": ["vite/client"],
+        "noImplicitReturns": false,
+        "target": "ES2019",
+        "moduleResolution": "Node"
+    }
+}
